move resourceMapper funcs to 'reconcile' package
